+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 * Audit fix of dependencies
 * Updated dependencies
 * Added support for contracting an abstract class
-* Fixed bug [#210](https://github.com/final-hill/decorator-contracts/issues/210)
+* Fixed bug [#210](https://github.com/final-hill/decorator-contracts/issues/210) for supporting subcontract private fields
+* Added `within` feature for declaring timing requirements
 
 ## v0.20.5